+++ src/AllowedFilter.php
@@ -4,9 +4,11 @@ namespace Spatie\QueryBuilder;
 
 use Illuminate\Support\Collection;
 use Spatie\QueryBuilder\Filters\Filter;
+use Spatie\QueryBuilder\Filters\FiltersCallback;
 use Spatie\QueryBuilder\Filters\FiltersExact;
-use Spatie\QueryBuilder\Filters\FiltersScope;
 use Spatie\QueryBuilder\Filters\FiltersPartial;
+use Spatie\QueryBuilder\Filters\FiltersScope;
+use Spatie\QueryBuilder\Filters\FiltersTrashed;
 
 class AllowedFilter
 {
@@ -44,32 +46,53 @@ class AllowedFilter
             return;
         }
 
-        ($this->filterClass)($query, $valueToFilter, $this->internalName);
+        ($this->filterClass)($query->getEloquentBuilder(), $valueToFilter, $this->internalName);
     }
 
-    public static function exact(
-        string $name,
-        ?string $internalName = null,
-        bool $addRelationConstraint = true
-    ): self {
+    public static function setFilterArrayValueDelimiter(string $delimiter = null): void
+    {
+        if (isset($delimiter)) {
+            QueryBuilderRequest::setFilterArrayValueDelimiter($delimiter);
+        }
+    }
+
+    public static function exact(string $name, ?string $internalName = null, bool $addRelationConstraint = true, string $arrayValueDelimiter = null): self
+    {
+        static::setFilterArrayValueDelimiter($arrayValueDelimiter);
+
         return new static($name, new FiltersExact($addRelationConstraint), $internalName);
     }
 
-    public static function partial(
-        string $name,
-        $internalName = null,
-        bool $addRelationConstraint = true
-    ): self {
+    public static function partial(string $name, $internalName = null, bool $addRelationConstraint = true, string $arrayValueDelimiter = null): self
+    {
+        static::setFilterArrayValueDelimiter($arrayValueDelimiter);
+
         return new static($name, new FiltersPartial($addRelationConstraint), $internalName);
     }
 
-    public static function scope(string $name, $internalName = null): self
+    public static function scope(string $name, $internalName = null, string $arrayValueDelimiter = null): self
     {
+        static::setFilterArrayValueDelimiter($arrayValueDelimiter);
+
         return new static($name, new FiltersScope(), $internalName);
     }
 
-    public static function custom(string $name, Filter $filterClass, $internalName = null): self
+    public static function callback(string $name, $callback, $internalName = null, string $arrayValueDelimiter = null): self
     {
+        static::setFilterArrayValueDelimiter($arrayValueDelimiter);
+
+        return new static($name, new FiltersCallback($callback), $internalName);
+    }
+
+    public static function trashed(string $name = 'trashed', $internalName = null): self
+    {
+        return new static($name, new FiltersTrashed(), $internalName);
+    }
+
+    public static function custom(string $name, Filter $filterClass, $internalName = null, string $arrayValueDelimiter = null): self
+    {
+        static::setFilterArrayValueDelimiter($arrayValueDelimiter);
+
         return new static($name, $filterClass, $internalName);
     }
 
@@ -122,7 +145,7 @@ class AllowedFilter
     protected function resolveValueForFiltering($value)
     {
         if (is_array($value)) {
-            $remainingProperties = array_diff($value, $this->ignored->toArray());
+            $remainingProperties = array_diff_assoc($value, $this->ignored->toArray());
 
             return ! empty($remainingProperties) ? $remainingProperties : null;
         }