+++ src/Includes/IncludedRelationship.php
@@ -2,11 +2,15 @@
 
 namespace Spatie\QueryBuilder\Includes;
 
-use Illuminate\Support\Collection;
+use Closure;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Support\Collection;
 
 class IncludedRelationship implements IncludeInterface
 {
+    /** @var Closure|null */
+    public $getRequestedFieldsForRelatedTable;
+
     public function __invoke(Builder $query, string $relationship)
     {
         $relatedTables = collect(explode('.', $relationship));
@@ -15,7 +19,9 @@ class IncludedRelationship implements IncludeInterface
             ->mapWithKeys(function ($table, $key) use ($query, $relatedTables) {
                 $fullRelationName = $relatedTables->slice(0, $key + 1)->implode('.');
 
-                $fields = $query->getRequestedFieldsForRelatedTable($fullRelationName);
+                if ($this->getRequestedFieldsForRelatedTable) {
+                    $fields = ($this->getRequestedFieldsForRelatedTable)($fullRelationName);
+                }
 
                 if (empty($fields)) {
                     return [$fullRelationName];
@@ -33,7 +39,7 @@ class IncludedRelationship implements IncludeInterface
     public static function getIndividualRelationshipPathsFromInclude(string $include): Collection
     {
         return collect(explode('.', $include))
-            ->reduce(function ($includes, $relationship) {
+            ->reduce(function (Collection $includes, string $relationship) {
                 if ($includes->isEmpty()) {
                     return $includes->push($relationship);
                 }