+++ tests/FieldsTest.php
@@ -4,12 +4,12 @@ namespace Spatie\QueryBuilder\Tests;
 
 use DB;
 use Illuminate\Http\Request;
-use Spatie\QueryBuilder\QueryBuilder;
-use Spatie\QueryBuilder\Exceptions\InvalidFieldQuery;
-use Spatie\QueryBuilder\Tests\TestClasses\Models\TestModel;
-use Spatie\QueryBuilder\Exceptions\UnknownIncludedFieldsQuery;
-use Spatie\QueryBuilder\Tests\TestClasses\Models\RelatedModel;
 use Spatie\QueryBuilder\Exceptions\AllowedFieldsMustBeCalledBeforeAllowedIncludes;
+use Spatie\QueryBuilder\Exceptions\InvalidFieldQuery;
+use Spatie\QueryBuilder\Exceptions\UnknownIncludedFieldsQuery;
+use Spatie\QueryBuilder\QueryBuilder;
+use Spatie\QueryBuilder\Tests\TestClasses\Models\RelatedModel;
+use Spatie\QueryBuilder\Tests\TestClasses\Models\TestModel;
 
 class FieldsTest extends TestCase
 {
@@ -45,6 +45,23 @@ class FieldsTest extends TestCase
         $expected = TestModel::query()->toSql();
 
         $this->assertEquals($expected, $query);
+    }
+
+    /** @test */
+    public function it_replaces_selected_columns_on_the_query()
+    {
+        $query = $this
+            ->createQueryFromFieldRequest(['test_models' => 'name,id'])
+            ->select(['id', 'is_visible'])
+            ->allowedFields(['name', 'id'])
+            ->toSql();
+
+        $expected = TestModel::query()
+            ->select("{$this->modelTableName}.name", "{$this->modelTableName}.id")
+            ->toSql();
+
+        $this->assertEquals($expected, $query);
+        $this->assertStringNotContainsString('is_visible', $expected);
     }
 
     /** @test */
@@ -131,8 +148,8 @@ class FieldsTest extends TestCase
 
         $queryBuilder->first()->relatedModels;
 
-        $this->assertQueryLogContains('select "test_models"."id" from "test_models"');
-        $this->assertQueryLogContains('select "name" from "related_models"');
+        $this->assertQueryLogContains('select `test_models`.`id` from `test_models`');
+        $this->assertQueryLogContains('select `name` from `related_models`');
     }
 
     /** @test */
@@ -222,8 +239,8 @@ class FieldsTest extends TestCase
 
         $queryBuilder->first()->relatedModels;
 
-        $this->assertQueryLogContains('select * from "test_models"');
-        $this->assertQueryLogContains('select "id", "name" from "related_models"');
+        $this->assertQueryLogContains('select * from `test_models`');
+        $this->assertQueryLogContains('select `id`, `name` from `related_models`');
     }
 
     /** @test */