+++ tests/FilterTest.php
@@ -2,16 +2,16 @@
 
 namespace Spatie\QueryBuilder\Tests;
 
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\Request;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\DB;
-use Spatie\QueryBuilder\QueryBuilder;
 use Spatie\QueryBuilder\AllowedFilter;
-use Illuminate\Database\Eloquent\Builder;
-use Spatie\QueryBuilder\Filters\FiltersExact;
 use Spatie\QueryBuilder\Exceptions\InvalidFilterQuery;
 use Spatie\QueryBuilder\Filters\Filter as CustomFilter;
 use Spatie\QueryBuilder\Filters\Filter as FilterInterface;
+use Spatie\QueryBuilder\Filters\FiltersExact;
+use Spatie\QueryBuilder\QueryBuilder;
 use Spatie\QueryBuilder\Tests\TestClasses\Models\TestModel;
 
 class FilterTest extends TestCase
@@ -32,6 +32,19 @@ class FilterTest extends TestCase
         $models = $this
             ->createQueryFromFilterRequest([
                 'name' => $this->models->first()->name,
+            ])
+            ->allowedFilters('name')
+            ->get();
+
+        $this->assertCount(1, $models);
+    }
+
+    /** @test */
+    public function it_can_filter_models_by_an_array_as_filter_value()
+    {
+        $models = $this
+            ->createQueryFromFilterRequest([
+                'name' => ['first' => $this->models->first()->name],
             ])
             ->allowedFilters('name')
             ->get();
@@ -94,7 +107,7 @@ class FilterTest extends TestCase
             ->toSql();
 
         $expectedSql = TestModel::select('id', 'name')
-            ->where(DB::raw('LOWER("name")'), 'LIKE', 'john')
+            ->where(DB::raw('LOWER(`name`)'), 'LIKE', 'john')
             ->toSql();
 
         $this->assertEquals($expectedSql, $queryBuilderSql);
@@ -188,6 +201,32 @@ class FilterTest extends TestCase
     }
 
     /** @test */
+    public function it_can_filter_results_by_type_hinted_scope()
+    {
+        TestModel::create(['name' => 'John Testing Doe']);
+
+        $modelsResult = $this
+            ->createQueryFromFilterRequest(['user' => 1])
+            ->allowedFilters(AllowedFilter::scope('user'))
+            ->get();
+
+        $this->assertCount(1, $modelsResult);
+    }
+
+    /** @test */
+    public function it_can_filter_results_by_regular_and_type_hinted_scope()
+    {
+        TestModel::create(['id' => 1000, 'name' => 'John Testing Doe']);
+
+        $modelsResult = $this
+            ->createQueryFromFilterRequest(['user_info' => ['id' => '1000', 'name' => 'John Testing Doe']])
+            ->allowedFilters(AllowedFilter::scope('user_info'))
+            ->get();
+
+        $this->assertCount(1, $modelsResult);
+    }
+
+    /** @test */
     public function it_can_filter_results_by_scope_with_multiple_parameters()
     {
         Carbon::setTestNow(Carbon::parse('2016-05-05'));
@@ -203,12 +242,27 @@ class FilterTest extends TestCase
     }
 
     /** @test */
+    public function it_can_filter_results_by_scope_with_multiple_parameters_in_an_associative_array()
+    {
+        Carbon::setTestNow(Carbon::parse('2016-05-05'));
+
+        $testModel = TestModel::create(['name' => 'John Testing Doe']);
+
+        $modelsResult = $this
+            ->createQueryFromFilterRequest(['created_between' => ['start' => '2016-01-01', 'end' => '2017-01-01']])
+            ->allowedFilters(AllowedFilter::scope('created_between'))
+            ->get();
+
+        $this->assertCount(1, $modelsResult);
+    }
+
+    /** @test */
     public function it_can_filter_results_by_a_custom_filter_class()
     {
         $testModel = $this->models->first();
 
         $filterClass = new class implements FilterInterface {
-            public function __invoke(Builder $query, $value, string $property) : Builder
+            public function __invoke(Builder $query, $value, string $property): Builder
             {
                 return $query->where('name', $value);
             }
@@ -421,12 +475,12 @@ class FilterTest extends TestCase
     /** @test */
     public function it_should_apply_a_default_filter_value_if_nothing_in_request()
     {
-        TestModel::create(['name' => 'John Doe']);
-        TestModel::create(['name' => 'John Deer']);
+        TestModel::create(['name' => 'UniqueJohn Doe']);
+        TestModel::create(['name' => 'UniqueJohn Deer']);
 
         $models = $this
             ->createQueryFromFilterRequest([])
-            ->allowedFilters(AllowedFilter::partial('name')->default('John'))
+            ->allowedFilters(AllowedFilter::partial('name')->default('UniqueJohn'))
             ->get();
 
         $this->assertEquals(2, $models->count());
@@ -435,14 +489,14 @@ class FilterTest extends TestCase
     /** @test */
     public function it_does_not_apply_default_filter_when_filter_exists_and_default_is_set()
     {
-        TestModel::create(['name' => 'John Doe']);
-        TestModel::create(['name' => 'John Deer']);
+        TestModel::create(['name' => 'UniqueJohn UniqueDoe']);
+        TestModel::create(['name' => 'UniqueJohn Deer']);
 
         $models = $this
             ->createQueryFromFilterRequest([
-                'name' => 'Doe',
+                'name' => 'UniqueDoe',
             ])
-            ->allowedFilters(AllowedFilter::partial('name')->default('John'))
+            ->allowedFilters(AllowedFilter::partial('name')->default('UniqueJohn'))
             ->get();
 
         $this->assertEquals(1, $models->count());
@@ -455,5 +509,39 @@ class FilterTest extends TestCase
         ]);
 
         return QueryBuilder::for(TestModel::class, $request);
+    }
+
+    /** @test */
+    public function it_can_override_the_array_value_delimiter_for_single_filters()
+    {
+        TestModel::create(['name' => '>XZII/Q1On']);
+        TestModel::create(['name' => 'h4S4MG3(+>azv4z/I<o>']);
+
+        // First use default delimiter
+        $models = $this
+            ->createQueryFromFilterRequest([
+                'ref_id' => 'h4S4MG3(+>azv4z/I<o>,>XZII/Q1On',
+            ])
+            ->allowedFilters(AllowedFilter::exact('ref_id', 'name', true))
+            ->get();
+        $this->assertEquals(2, $models->count());
+
+        // Custom delimiter
+        $models = $this
+            ->createQueryFromFilterRequest([
+                'ref_id' => 'h4S4MG3(+>azv4z/I<o>|>XZII/Q1On',
+            ])
+            ->allowedFilters(AllowedFilter::exact('ref_id', 'name', true, '|'))
+            ->get();
+        $this->assertEquals(2, $models->count());
+
+        // Custom delimiter, but default in request
+        $models = $this
+            ->createQueryFromFilterRequest([
+                'ref_id' => 'h4S4MG3(+>azv4z/I<o>,>XZII/Q1On',
+            ])
+            ->allowedFilters(AllowedFilter::exact('ref_id', 'name', true, '|'))
+            ->get();
+        $this->assertEquals(0, $models->count());
     }
 }